+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.22.1
+
+- Fixed old `pb.authStore.isAdmin`/`pb.authStore..isAuthRecord` and marked them as depricated in favour of `pb.authStore.isSuperuser` ([#323](https://github.com/pocketbase/js-sdk/issues/323)).
+    _Note that with PocketBase v0.23.0 superusers are converted to a system auth collection so you can always simply check the value of `pb.authStore.record?.collectionName`._
+
+
 ## 0.22.0
 
 **⚠️ This release introduces some breaking changes and works only with PocketBase v0.23.0+.**